Extract references from raw ReSpec documents (#1)
The main difficulty was to have jsdom run ReSpec and, since ReSpec runs
asynchronously, to get a call back when ReSpec is done with the generation
of the spec. I attached to the "postProcess" event of ReSpec which has the
advantage of being an array (so I won't override post-processing that the
spec author may have put in place), and the drawback of running before
"afterEnd", which means I may miss some of the post-processing. That
post-processing is very unlikely to ever affect references, though.

I note that both "postProcess" and "afterEnd" events are considered deprecated
in ReSpec and may be removed in a future version. Not sure how we could plug
into ReSpec afterwards (can we plug into the pubsubhub system from external
code?)