Cache folder created and reset as needed
The fetch function now automatically creates a "cache" folder to store its
results unless that folder already exists. The contents of that folder are
reset when the program starts unless the "preserveCache" flag is set in
"config.json"

All HTTP requests are stored, including requests to api.w3.org. Note that
since we pass the API key directly as a querystring parameter for the time
being, this means that the key appears in clear in the corresponding filename
in the cache folder. That is not optimal.