W3C API key now passed in Authorization header
The key used to be passed as a querystring parameter but this meant that it
appeared in clear in the name of the response file cached in the cache
subfolder, which was not particularly "clean".

Note that the "fetch" method defined in util.js is totally ignorant of the HTTP
headers that may be sent along with the request. It matches cached entries on
the URL alone.