Fetch now sends conditional HTTP requests when possible (#21)
The fetch function now starts by looking in the cache folder for previous
response headers. If found and if these headers contain a Last-Modified and/or
Etag directive, fetch sets the appropriate "If-Modified-Since" and
"If-None-Match" HTTP request headers, and reuses the local version of the
response upon receipt of a 304.

Note that this update also reinstates chunk processing for the list of specs
to avoid "socket hang up" messages that arose from time to time, likely due to
too many requests sent to a given domain name (not sure how Node.js pools such
requests internally). The crawler currently parses 10 specs at a time.

Possible new parameters in config.json:
- avoidNetworkRequests: if true, the fetch function will directly use the
contents of the cache for the requested URL, if it exists
- resetCache: if true, the program empties the cache folder when it starts

The "preserveCache" flag is no longer supported (the program now preserves the
cache by default and that is a good thing).