Normalized initial list of spec URLs
Now always using the shortname version of the URL when available and possible.
Also updated the link to the Media Sessions spec which transitioned from WHATWG
to WICG.

This update also upgrades the version of jsdom to 9.4.1 to get DOMParser
support. This does not yet change anything in practice until bug with
"instanceof Function" gets solved on ReSpec side.