IDL parser now exports "callback" functions (#24)
The IDL parser now correctly includes callback functions in the `idlNames`
property.

This change means that the `localNames` property becomes useless: all IDL names
defined in a spec are to be exported. I dropped all mentions of `localNames`
from the code as a result.

This commit also introduces the first few tests for the IDL parser. Tests use
Jasmine. These first basic tests typically ensure that the IDL parser exports
the IDL names as expected.