Primary global interface name now based on [PrimaryGlobal] attr (#25)
The IDL parser now correctly detects the interface identified with a
`[PrimaryGlobal]` extended attribute, and correctly links interfaces
defined without `[Exposed]` to it in the end.

If no `[PrimaryGlobal]` interface is found in the IDL content, the IDL parser
continues to assume that the primary global interface is `Window` as before,
in other words that we are in a Web context.