Use respec to generate DOM from respec documents (#31)
The code now uses ReSpec doc writer to retrieve the DOM of a ReSpec document. Note that this means the document gets rendered in a real browser in a virtual display through Nightmare.js, which seems a tad overkill compared to using jsdom, but needed since jsdom does not support all JS APIs that ReSpec uses (jsdom typically does not support URL.createObjectURL for the time being). Using a real browser is more future-proof in that regard.

The update upgrades jsdom to latest version.

Also, rendering errors no longer crash the crawl. Errors are now reported within an `error` property in the JSON report that the crawl generates.

Note that using ReSpec doc writer for rendering means that the tool no longer runs offline, even when it has everything in cache.