Re-write find-spec as module, make report a param
The path to the current crawl report should now be passed as parameter
(not a good idea to have it hardcoded given that this report is not the Git
repository).

If needed, find-spec could now be used from another Node.js file.

Also, this commit slightly improves the readability of the code.