Improve references extraction rules to skip over early "References" sections
In the absence of a proper spec generator, the code looks for "References"
sections. It used to lose itself on the HTML LS spec because that spec has a
"References" section that has nothing to do with the actual list of references.

Rules have been improved a bit to handle that case. That's still not perfect,
but new rules seems to work as before on all other specs and now correctly
extract references from the HTML LS spec.