Flag specs with obsolete WebIDL Level 1 constructs in crawl report
Some specs still use `[]` instead of `FrozenArray`. Our WebIDL parser
automatically updates these definitions to be able to process the WebIDL.

The fact that this had to be done is now reported in an `hasObsoleteIDL`
property.

Note that this property is just a boolean for now. Ideally, it would list the
exact IDL definitions that need to be fixed.