"Import" mechanism for list of specs to reference a common list
To ease maintenance and better understand which specs are different from a W3C
perspective and from a WHATWG perspective, the list of specs that gets passed
to the crawler can now reference another list of specs defined in another file.

This allowed the creation of a specs-common.json file now referenced by both
specs-w3c.json and specs-whatwg.json.

Note that mechanism is not recursive for now and will have to be adjusted if we
need more structure later on.