First stab at dependencies report
The study tool can now generate a dependencies report when parameter is "dep".
That dependencies report lists specs that got crawled. For each of them, the
report lists the specs that reference it either normatively or informatively.

The report could be further improved. It could for instance clarify what
version of the spec the reference targets (ED, version number when available,
dated spec, generic TR spec, etc.). It could also list the references that are
currently missing but that should exist (one of the anomalies that we report
in other reports).