Improve missing/inconsistent ref reporting
The crawler now skips over the "head" section of W3C specs when it extracts
links, because that section typically contains links to documents that do not
need to appear in the list of references.

The study tool now also ignores links to self, including links to other
versions of self. As pointed out in a code comment, there may be cases where
it would actually be worth reporting these links as potentially problematic,
but most of the times they simply appear in changelog sections.