Report ref URLs that are inconsistent with URLs used in links (#49)
Thanks to equivalence and crawling, we can now mostly easily map between
different versions of the same spec (ED, dated spec, latest published version,
previous version, etc.). However, while this mapping helps spotting missing
references to specs, doing it automatically can also hide possible
inconsistencies in a spec when the body of the spec may link to a particular
version of a spec whereas the final reference is to another version.

Reports now include a section that alerts about such inconsistencies.

Internally, canonicalization now has two levels:
1. by default, it returns the "same" document
2. if requested, it may handle dated specs and take equivalences into account


The crawler now also skips over the "head" section of W3C specs when it extracts
links, because that section typically contains links to documents that do not
need to appear in the list of references.

The study tool now also ignores links to self, including links to other
versions of self. As pointed out in a code comment, there may be cases where
it would actually be worth reporting these links as potentially problematic,
but most of the times they simply appear in changelog sections.

The list of specs and equivalences was completed to remove false positives.