First stab at producing an HTML version of the report per spec
The HTML version of the report gets generated from the Markdown version
using Pandoc. It includes a bit of JavaScript that automatically collapses
sections and adds a few icons to report on errors.

The whole thing is pretty rough for the time being. Filtering should be added,
and styles improved.

The commit also adds conveniences Makefile-like scripts to package.json to
produce the reports.