Updated report generation scripts to ease automation
The reports are now generated in a reports/w3c (or reports/whatwg) folder.
Files in that folder can directly be copied over to the corresponding folder in
the "reffy-reports" repository.

Common stylesheets and JS files for the HTML report moved to the Reffy reports
repository as well.

Note the generated reports should NOT be checked in the Reffy repository, which
should only contain Reffy's code.