Report and render network errors consistently
The crawler ignored errors reported by the W3C API and simply fetched the latest
published version instead of the Editor's Draft when such an error occurred
(the crawler uses the W3C API to get the URL of the Editor's Draft).

These errors are now reported in the crawl data file and rendered in the final
Markdown and HTML report.

Also, results in the crawl data file are now sorted by spec URL to ease the
detection of diffs across crawls.