Add date to crawl report (#63)
The crawler now adds the crawling date to the JSON results. It also stores the
number of specifications that were crawled and the number of specifications
for which the crawl ended up with an error.

This update modifies the structure of the JSON file. The first level is no
longer an array in particular: it is an object such as:

{
  "date": "2017-07-03T20:49:21.821Z",
  "stats": {
    "crawled": 113,
    "errors": 1
  },
  "results" [...]
}

The crawl date is now also added to the Markdown and HTML versions of the
reports. The study tool and the conversion to HTML also supports the notion
of a crawl title (see #64), but note that the crawler does not yet append any
title to the crawl result.