Extend canonicalization rules to handle variants
The same URL can be canonicalized into different possibilities. For instance,
Canvas 2D, Workers, Web Messaging, Web Sockets can all be considered as
alternatives to the WHATWG HTML spec which includes them. These alternatives
need to be taken into account to correctly report inconsistent references
instead of missing references.

For instance, Reffy used to report a missing reference to HTML 5.2 because
the canonicalization of the WHATWG HTML spec actually returned the URL of the
Workers spec at W3C (one of the actual possibilities).

`canonicalizeURL` can now return an array of equally valid canonicalized URL
if so requested with the `returnAlternatives` option, and the module now also
exports a `canonicalizesTo` method that returns true when a given URL could
be canonicalized as a reference URL or as one URL in a list.