Add title/description to list of input specifications (#64)
The list of specs given as input to the crawler can now be an JSON object
with a `title` and `description` properties that get preserved and output
to the data crawl structure.

The title is also used as title for the Markdown and HTML reports generated
afterwards. Note the description is not used for now (no easy way to add
custom variables in Markdown for pandoc).