Add URLs to latest published version and ED to reports
The reports only used to mention the URL that was crawled which, depending on
configuration options, could have been the Editor's Draft or the latest
published version.

Even when the latest published version gets crawled, it is still handy to have
a link to the ED.

The crawl report is now more complete about the URLs associated with each spec.
It contains the following spec properties:

- url: the initial URL, as it appears in the initial JSON file
- latest: the "Latest Published Version" link (http://www.w3.org/TR/shortname/)
- datedUrl: the dated URL of the latest published version
- datedStatus: the latest publication status (e.g. Recommendation)
- edDraft: the URL of the Editor's Draft
- crawled: the URL that was crawled

The URL that gets crawled is either:
- the datedUrl (or the latest URL if the dated URL could not be found for some
reason) when the "publishedVersion" crawl option is on, or when the ED is known
to be bogus
- the Editor's Draft otherwise, when it's known
- or the initial URL (for WHATWG specs, typically)

These URLs are highlighted in the report in the "Spec info" part.

(This update was suggested in #74)