Ignore "Window" in "Exposed=Window" for missing references
As described in #72, WebIDL now requires explicit `Exposed=Window` statements,
and requiring a normative reference to HTML for all specs that define IDL terms
seems a bit overkill. The idea is thus to treat that as an exception to the
rule.

In practice, Reffy's IDL parser tracks references to `Window` and sets an
internal `_reallyDependsOnWindow` flag in the results of the crawl when the
IDL really depends on `Window` (meaning when `Window` appears in statements
other than `Exposed=Window`).

The report generator then uses that information to ignore `Window` dependencies
as needed.

This updates makes the Ambient Light Sensor spec turn "all green" for instance.