Fix "No IDL content" logic, broken by previous commit
Last commit introduced a `_reallyDependsOnWindow` internal key in the `idl`
object to create an exception to the rule for `Window`.

However, the logic that set the "No IDL content" flag relied on the number of
keys in the `idl` object, assuming there would be only one internal key. This
was missed in previous commit because that logic was not explicit that it was
taking internal keys into account. The logic was fixed and is now explicit.