Retry network fetches a few times to work around transient errors
For some reason, fetching the SVG 2 spec fails with a "socket hang up" from
time to time. The fetch function now tries again after a few seconds when
such errors happen. Hopefully, that will make Reffy a bit more resilient to
transient network issues.

Not related, but the result of the check on pending fetches was actually not
used, meaning Reffy could fetch the same URL multiple times at once. Fixed in
this commit as well.