New checker and merger tools, study tool split, HTML input for crawler
This commit makes a number of updates to the code in order to create a
spec checker tool that takes the URL of a spec as input and analyzes the
underlying spec against a reference crawl report to produce an anomalies report
for that particular spec.

These updates (and the new spec checker tool) also provide the foundations for
the HTTP service described in #76.

Main changes are:
- The crawler now accepts an object with an `html` property that contains the
HTML of the specification to crawl. Note that the crawler will fail to crawl
the source of a ReSpec document though, because it needs to call `RespecWriter`
which expects a URL.
- A new crawl results merger tool was created. It merges a crawl report into
a reference one to produce a new crawl report. This allows incremental updates
to existing reports, which in turn is useful to study a particular spec.
- The crawl report analysis was moved out of the old study tool and into its
own new module/tool named `study-crawl.js`. That tool returns a JSON report for
the list of specs that were crawled, or for a given spec or set of specs.
- The old study tool was renamed into `generate-report.js`. It does the exact
same thing as before, generating human-readable reports. However, note it can
now take either the results of a crawl as input or a study of these results,
generated by the `study-crawl.js` tool. Internally, if the tool receives the
results of a crawl as input, it will simply start by calling the
`study-crawl.js` tool.
- A new `check-specs.js` tool was created to crawl and study a given spec (or
list of specs) against a reference crawl results. Essentially, it takes the URL
of the spec to check, a reference crawl, and the name of the resulting file to
create as inputs, and applies the crawler, the merger and the new study tool in
order. This produces the anomalies report for the given spec. Once it exists,
the HTTP service would typically call the `checkSpecs` method exposed by this
module.
- Script commands now generate `study.json` files that contain the analysis in
JSON. That is not stricto senso required, but it seems useful to keep track of
that step too.
- The README was updated to mention the new tools and update the description of
tools that were changed.
- Each tool now starts with some JSDoc that describes what it does.