Bump dependencies versions (to fix ReSpec writer crash)
Calls to ReSpec doc writer no longer worked with the online version of ReSpec
for some reason, resulting in timeouts when trying to extract info from
ReSpec-based specs. Bumping the version of ReSpec fixes everything.

Also took the occasion to bump the version of jsdom, node-fetch, and of the
WebIDL parser.