Turn second parameter of crawler into a folder
The crawler exports IDL fragments for each spec in separate `idl` files on
top of the crawl report. It did not make sense to request the user to provide
the name of the crawl report and to expect the user to have created an `idl`
folder.

The crawler now expects the second parameter to be the name of an existing
folder into which it will write the crawl report and create an `idl` folder
that contains the IDL exports.