New reffy.js CLI, fix path issues
To run as an NPM package, tools need to handle all relative paths to files and
folders provided by the user as relative to the current working directory, and
not as relative to the directory of the JS file. That was already the case for
file operations using `fs`, but `require` resolves path relative to the JS file
and thus cannot be used directly to parse JSON files.

Also, scripts defined in `package.json` cannot easily be run when Reffy is
downloaded as an NPM package. The new `reffy.js` CLI is meant to address this.
It contains the logic that was previously in these scripts. Some tools had to
be adjusted as a result, notably the report generation tool.

README file updated to mention the NPM package.

Note: update also includes a fix on the Khronos URL normalization logic. A typo
there was crashing the crawler.