Use fetch-filecache-for-crawling npm package
The "fetch" function defined in util.js that makes Reffy use a local cache
folder was moved to a separate npm package named fetch-filecache-for-crawling
so that it can be referenced by other projects (notably w3c/web-roadmaps).

This update replaces the definition of the fetch function by a reference to the
npm package. This does not trigger any change, but note that:
1. the default cache folder name is now `.cache` (it was `cache`). README file
updated accordingly.
2. fetch logs are now disabled by default. To enable them, one needs to set a
`logToConsole` property into `config.json`.