Bump version of fetch package, read parameters from config.json
Switched to latest version of fetch-filecache-for-crawling that introduces a
new `setParameter` function and stops reading these parameters from
`config.json`.

As a result, Reffy now reads these parameters from `config.json` directly and
calls `setParameter` as needed.

Note the `avoidNetworkRequests` and `forceRefresh` parameters are no longer
supported. To be replaced in the config file by a parameter named `cacheRefresh`
that can take any of the values of the `refresh` parameter described in:
https://github.com/tidoust/fetch-filecache-for-crawling#configuration

Also, note the fetch package now defaults to behaving like a regular HTTP cache
whereas we typically want Reffy to check all URLs at least once when we run a
crawl. Reffy sets the `refresh` parameter to `once` accordingly for fetch
requests, unless otherwise specified in the `config.json` file.