Complete list of specs to crawl
Goal is to converge on a common list of specs with Web Platform Tests to
generate the right IDL files. See:
https://github.com/tidoust/reffy-reports/issues/3#issuecomment-387389425

This update creates a new `specs-cg.json` file that contains the list of specs
that define some IDL content incubated in the WICG or in other Community Groups.
That file is referenced from the W3C / WHATWG crawls.

This update also adds specs that were still missing.