Warn when W3C spec does not link to an Editor's Draft
The analysis now reports a warning when the W3C API did not return the URL of
an Editor's Draft. In turn, this should mean that the spec does not include
such a link, or that it does not include it along with the "this version" et al.
links.

Fixes #104.