Have ReSpec run in JSDOM, stop using RespecDocWriter
Reffy needs to run ReSpec on Editor's Drafts. It relied on RespecDocWriter
until now, which uses Puppeeter, which in turn relies on a headless version
of Chrome. That requires downloading 100s of MB of dependencies and is a tad
heavy for the task at hands, especially as we consider running Reffy as an
HTTP service.

With this update, RespecDocWriter is no longer being used. Instead, Reffy makes
ReSpec run in JSDOM, which is much more lightweight (and JS-based).

Problem is that JSDOM does not yet support enough features to run ReSpec.
Concrete issues are documented in:
https://github.com/w3c/respec/issues/1469#issuecomment-388560878

This update monkey-patches JSDOM's code to add good-enough implementations of
missing features, as well as to regain control over network requests so that
all network requests go through our cache. It also monkey-patches ReSpec to:

1. drop a couple of modules that cannot run because they use non-implemented
features of JSDOM that cannot be easily patched
2. update the code of a couple of ReSpec dependencies for the same reason

The whole thing remains fragile: updates to ReSpec or to its dependencies may
well break the regular expressions used to monkey patch the code. Most of the
monkey-patching should resist minor changes. A couple of them may not though.