[Crawl] Crawl each spec in a separate child process
The crawl of a given spec can be memory and CPU intensive. For instance, the
crawl of the HTML spec takes about 1Gb of memory. Running each of them in a
separate process guarantees isolation and freeing of resources as soon as
possible. From a debugging perspective, it also makes it easier to track a
particular spec processing within the task manager.

The crawler spawns 10 child processes at once at most (the number is hardcoded
for now, but could be turned into a configuration setting if needed). This
mechanism replaces the previous processing in chunks, and improves it as well
since new specs get processed as soon as a new child process can be created.

The crawler now monitors and kills child processes that take too long (more
than one minute, hardcoded as well for now).

For what it's worth, one consequence of the change is that the `crawl.json` is
only written at the end of the crawl (and not filled out progressively after a
chunk has been processed).

Performance wise, in the absence of network requests, a crawl used to take
about 100s. It now takes 60s.