[Crawl] Make parent process handle all fetch requests
Reffy crawls each spec in a separate process. Problem is it also uses the
`fetch-filecache-for-crawling` library to use a file cache for `fetch` requests
and that library is not thread-safe. Notably, if multiple processes attempt to
read/write the same file at once (which typically happens for fetch requests
on ReSpec), reads may be corrupted, leading to apparently random errors
(although always on ReSpec specs in practice, because that's the only resource
we need to fetch multiple times in practice).

Making the library thread-safe would be ideal, but requires using file locks
which I'm not crazy about. In the meantime, this update makes the child process
send `fetch` requests to its parent process so that all requests to the library
get done by the parent process, effectively removing multi-threading issues.