Add CSS specs to crawl and make crawler CSS/IDL-aware
This update adds about 85 CSS specifications to the list of crawled specs. It
also separates the specs into three main buckets:
- those that define IDL content: in filenames that end with `-idl`
- those that define CSS content: in filenames that end with `-css`
- other specs: in `specs-other.json`

The crawler flags the specs internally as needed and the study tool only reports
IDL warnings for specs that are supposed to contain some IDL. The study tool
now also reports a warning if it finds some IDL content in a spec that is not
supposed to contain any.

Also, the CSS 2.1/2.2 specs are assembled manually from the multiple pages they
contain.