Add CSS definitions extractor
The extractor parses a CSS specification and extracts CSS definitions it
contains. The extractor currently extracts properties, descriptors and
non-terminal values (a.k.a. value spaces).

The extractor implements a number of ad-hoc extraction rules because CSS specs
use different conventions in practice. From time to time, the extractor finds
a non-terminal value definition but cannot find a production rule. In such
cases, it returns the prose that describes the definition.

Some non-terminal value definitions are only defined in prose although they
could also be defined with a real production rule. The `missing-css-rules.json`
file contains identified missing rules. It is not used anywhere for the time
being.

The CSS extracted from each spec is dumped in "pseudo" CSS Value Definition
Syntax in a `css` subfolder, where "pseudo" means that the dump contains a
couple of new syntaxes, namely:
- `property-name = foo` to describe the production rule of a property or of a
descriptor
- `property-name |= foo` to describe partials ("New values")

Only the most recent level gets dumped in that folder when the crawl contains
more than one level.