Apply all known patterns to extract IDL content
The IDL extractor assumed that a given spec would only use one and only one
pattern to define IDL content (e.g. `<pre><code class="idl">`). It tried
possible patterns in turn and used the first one that returned a match as a
result. This was done to avoid extracting other types of content by mistake.

Problem is the the Animation frames section in the HTML spec now uses a
different pattern (`<pre class="idl">` instead of `<pre><code class="idl">`),
which bugged the IDL extractor, as raised in:
https://github.com/tidoust/reffy/issues/125

The IDL extractor no longer stops when it has a match but rather applies all
known patterns. This fixes the problem with the HTML spec and does not seem to
create any issue with other specs.