Fix ReSpec document detection when `const` is used
Reffy uses the presence of a global `respecConfig` object as an indicator
that the underlying spec uses ReSpec. However, that object can be defined
with `const` and in that case, it does not appear under the `window` object
returned by JSDOM.

The code now uses `eval` (sigh!) to detect the existence of `respecConfig`.

The update also simplifies the `getDocumentAndGenerator` function not to wait
for ReSpec generation to be over (that's not needed, the code that loads the
document already waits!)