Don't try to send message when child process has exited
Page loading running in a child process may trigger a fetch in the background.
When that happens, Reffy may consider that the page has been loaded before the
fetch is done, and may thus proceed with the generated content and kill the
child process. Attempting to send the result of the fetch afterwards yields
an `ERR_IPC_CHANNEL_CLOSED`.

Parent process now only sends message when the child process is still around.