Extract IDL definitions from main spec sections, not from IDL index
Reffy gladly used the IDL index at the end of the spec to extract the IDL
definitions, but ReSpec generates this IDL index on-the-fly and we're currently
stuck to an old version of ReSpec (see #134), which chokes on recent WebIDL
constructs and fails to include them in the IDL index, as described in:
https://github.com/tidoust/reffy-reports/issues/34

With this update, Reffy now ignores the IDL index and extracts IDL definitions
from main spec sections, and not from the IDL index appendix.

Useless leading spaces are now discarded as well (#130).