WebIDL parsing updates to better support Global/Exposed/constructor (#186)
Introduces the following changes:
- Support for the new "constructor operation" syntax (#184). The old
constructor syntax through extended attributes is still supported
- Drop support for `[PrimaryGlobal]` and associated logic. That no longer exists
in WebIDL (this change means that the /TR/ crawl may report a few additional
hiccups, but so be it)
- Expose defined globals in a `globals` property. They used to be reported
under `idlNames` but they are neither JS names nor real IDL names, and this
created false positives when specs reuse the same global name on purpose to
inherit from other interfaces defined on it (for instance, the
`AnimationWorkletGlobalScope` interface reuses `Worklet`)
- Support for Global/Exposed attributes in partial definitions (#185)

Also, the update flattens the IDL parts of the report and in particular
drops `idlNames._dependencies` in favor of `dependencies`, and
`idlNames._reallyDependsOnWindow` in favor or `reallyDependsOnWindow`. Resulting
structure:

```json
{
  jsNames: {
      constructors: {},
      functions: {}
  },
  idlNames: {},
  idlExtendedNames: {},
  globals: {},
  dependencies: {},
  externalDependencies: [],
  reallyDependsOnWindow: false
}
```

The update fixes the incorrect reports on IDL interfaces being defined in
multiple specs (#115) as well.