Get back to RespecDocWriter to generate specs
Making latest version of Respec run under JSDOM is too much of a maintenance
burden in the end:
https://github.com/w3c/respec/issues/1469#issuecomment-546790835

This update makes Reffy use respecDocWriter (which uses the headless browser
Puppeteer under the hoods) again. This allows to run the latest version of
Respec.

JSDOM is still used to load specifications that do not use Respec and to detect
when a specification uses Respec. Note the code no longer runs external scripts
when it loads specifications: that seems unneeded in practice.

The code still contains minimal monkey patching for JSDOM so because a number
of specifications do have inline scripts that call `window.matchMedia`, which
JSDOM does not support.

This will fix #134.