Manage [Global] and [Exposed] names separately from IDL terms
See #194. The list of `[Global]` names referenced by a given spec through
`[Exposed]` extended attributes now appear in a separate `exposed` property,
along with the list of interfaces that reference each one.

Also, the list of `[Exposed]` names no longer appears under `dependencies` and
`externalDependencies`, which seems correct.

This update also:
- adds a new "Unknown [Exposed] names" anomaly to the report for names that
don't match any `[Global]` name;
- drops the `reallyDependsOnWindow` flag, since it's no longer needed;
- expects `[Global]` to take an identifier or an identifier list per the
WebIDL spec, and in particular no longer falls back to the name of the
interface.

As a result of the change, the report will no longer complain that a spec that
uses an `[Exposed]` name needs a normative reference to the spec that defines
the corresponding `[Global]`. Actually, I believe that is the right behavior,
because here may be more than one spec that defines a given `[Global]` name
(and that is the case in practice for `Worker` and `Worklet`). Typically, a
specs should not normatively reference the CSS Animation Worklet API just
because it defines an interface with `[Exposed=Worklet]`.