Make sure there can be only one CSS/IDL export per shortname
A crawl may contain different levels of the same spec. CSS/IDL exports are per
shortname and do not include any notion of level. The code that detected the
level of a given spec to only export the latest one could fail to detect that
this situation was happening. In turn, this could corrupt export files because
the code writes the exports concurrently.

This update makes sure that the crawler picks up only one spec per shortname,
and the latest version (or the first one in the crawl if we end up with two
specs with the same shortname and level for some reason).

Fixes #219.