Fix bug that made Reffy crawl the Compatibility standard 4 times
The "current position" counter was only incremented *after* a spec had been
processed. This meant that the same spec could be processed more than once.
The first spec in the list - the Compatibility Standard in practice - was
crawled 4 times in particular (4 being the number of processes that run in
parallel).

Fixes #233.