Track and report final redirected URL (#238)
In the crawl results, the `crawled` URL was the URL initially fetched by the
crawler, and not the final URL after possible redirects. With this update,
the URL that appears in `crawled` becomes the final redirected URL.

Note the `edDraft` URL remains untouched and is the URL of the Editor's Draft
as advertised in the published version of the spec.

See discussion in #220.