Fix detection of invalid WebIDL content in study tool (#247)
The structure of the error message returned by the WebIDL parser has changed
some time ago. It now features a `bareMessage` property, and not a `message`
property. The study tool was using the presence of a `message` property to
detect WebIDL error and set the "Invalid WebIDL content detected" flag.

Code updated to rather look for a `bareMessage` property.