Integrate feedback and add "delta" spec option
Code works well but latest level of some CSS specs can be defined as a delta
spec and the CSS and IDL that we want to extract is probably the one from the
non-delta spec for now. A spec can now be defined with a "delta" flag to alert
the crawler that no CSS/IDL extract file should be generated out of it.

A list of specs whose filename ends with "-delta" is considered to only contain
delta specs.